+++ .circleci/config.yml
@@ -8,7 +8,9 @@ jobs:
       - run: rustup toolchain install nightly
       - run: cargo +nightly -Zscript ./pre-publish.rs
       - run: cargo test
+      - run: cargo test --no-default-feature
       - run: cargo test --test 'serde' --features serde
+      - run: cargo test --test 'serde' --no-default-feature --features serde
   build_and_test_arm:
     machine:
       image: ubuntu-2004:current 
@@ -19,7 +21,9 @@ jobs:
       - run: rustup toolchain install nightly
       - run: cargo +nightly -Zscript ./pre-publish.rs
       - run: cargo test
+      - run: cargo test --no-default-feature
       - run: cargo test --test 'serde' --features serde
+      - run: cargo test --test 'serde' --no-default-feature --features serde
 
 workflows:
   version: 2
