+++ .github/workflows/tests.yml
@@ -32,8 +32,12 @@ jobs:
           cargo build
       - name: Run tests
         run: cargo test --verbose
+      - name: Run no-std tests
+        run: cargo test --verbose --no-default-feature
       - name: Run serde tests
         run: cargo test --verbose --test 'serde' --features serde
+      - name: Run no-std+serde tests
+        run: cargo test --verbose --test 'serde' --no-default-feature --features serde
 
   clippy:
     runs-on: ubuntu-latest
