+++ README.md
@@ -111,7 +111,8 @@ Any other overhead is considered a bug and should be reported.
 
 # Features
 
-- `serde`: implements `Serialize` and `Deserialize` for all types
+- `std`: enabled by default, required by some `f32` and `f64` methods.
+- `serde`: implements `Serialize` and `Deserialize` for all types.
 
 ## How it works
