+++ typed_floats/Cargo.toml
@@ -7,7 +7,7 @@ include = [
   "src/*.rs",
   "*.md",
 ]
-keywords = ["floats", "non-zero", "NaN"]
+keywords = ["floats", "non-zero", "NaN", "no_std"]
 license = "MIT OR Apache-2.0"
 name = "typed_floats"
 readme = "./README.md"
@@ -16,13 +16,22 @@ version = "0.3.0"
 rust-version = "1.56"
 
 [features]
-default = []
+default = ["std"]
 serde = ["dep:serde"]
+std = ["typed_floats_macros/std"]
 
 [dependencies]
-thiserror = "1.0"
-typed_floats_macros = {version = "=0.3.0", path = "../typed_floats_macros"}
-serde = { version = "1.0", features = ["derive"], optional = true }
+
+[dependencies.typed_floats_macros]
+version = "=0.3.0"
+path = "../typed_floats_macros"
+
+[dependencies.serde]
+version = "1.0"
+default-features = false
+features = ["derive"]
+optional = true
+
 
 [dev-dependencies]
 # For examples
