+++ typed_floats/tests/div_euclid.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "std")]
+
 use typed_floats::*;
 
 typed_floats_macros::generate_tests_self_rhs!(div_euclid);
