+++ typed_floats/tests/exp_m1.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "std")]
+
 use typed_floats::*;
 
 typed_floats_macros::generate_tests_self!(exp_m1);
