+++ typed_floats/tests/sin.rs
@@ -1,3 +1,5 @@
+#![cfg(feature = "std")]
+
 use typed_floats::*;
 
 typed_floats_macros::generate_tests_self!(sin);
