+++ typed_floats_macros/Cargo.toml
@@ -10,6 +10,9 @@ readme = "./README.md"
 [lib]
 proc-macro = true
 
+[features]
+std = []
+
 [dependencies]
 quote = "1.0"
 syn = { version = "2.0" }
